pyproject.toml
[tool.poetry]
name = "ansible-mysql-dynamic-inventory"
version = "0.1.0"
description = ""
authors = ["Larry Smith Jr. <mrlesmithjr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
ansible = "^4.8.0"
mysql-connector = "^2.2.9"

[tool.poetry.dev-dependencies]
bandit = "^1.7.1"
black = "^21.11b1"
flake8 = "^4.0.1"
molecule = {extras = ["docker"], version = "^3.5.2"}
mypy = "^0.910"
poetry = "^1.1.11"
pylint = "^2.11.1"
ansible-lint = "^5.2.1"
yamllint = "^1.26.3"
mkdocs = "^1.2.3"
autopep8 = "^1.6.0"
pycodestyle = "^2.8.0"
pip-tools = "^6.4.0"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
